pyproject.toml
[project]
name = "real-estate-scout"
version = "0.1.0"
description = "Simple CrewAI-powered real estate scout using LoopNet API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.28.0",
    "httpx>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "tenacity>=8.2.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "pytest>=8.0.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
